requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
httpie==2.3.0
idna==2.10
Markdown==3.3.3
Pillow==8.1.0
psycopg2==2.8.6
Pygments==2.7.4
PySocks==1.7.1
pytz==2020.5
qrcode==6.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
